#900 Combined tiled benchmark samples into one
